Fix editor opening wrong file in tree view mode
selected_file() was incorrectly using selected_file_index as a direct
index into the flat files list. In tree view mode, this index is the
visual position in the flattened tree, not the flat list index. Now
uses selected_tree_node() to get the correct file when in tree mode.